Extract admin-only ChannelType
